requirements.txt
numpy
opencv-contrib-python
openpyxl
pandas
Pillow
pyttsx3
tk



